+++ memo/ChangeLog.md
@@ -194,6 +194,7 @@
 - cmap: recognize <kbd>SS3 O</kbd> as <kbd>blur</kbd> `#D1384` 445a5ad
 - edit (`ble/widget/{accept-line,newline}`): automatically switch widgets by the keymap `#D1391` 5bed6e6
 - complete: perform filter in `ble/complete/cand/yield` `#D1404` 0000000
+- syntax, edit: use `type -a -t -- cmd` to get command types hidden by keywords `#D1406` 0000000
 
 <!---------------------------------------------------------------------------->
 # ble-0.4.0-devel1